Revert "Another, potentially bad perf optimisation that i might revert or keep idk"
This reverts commit b429da7c3c58cf0f60c03f11a95e7ad8e7c5c50f.
