--- packages/serialize/src/state.js
@@ -1,6 +1,0 @@
-// @flow
-import { type SerializedStyles } from '@emotion/utils'
-
-export let cursor: { current: void | SerializedStyles } = {
-  current: undefined
-}
